==> aou_workbench_pooled_analyses/03_merge_variants.ipynb <==
# Merge variants

In this notebook, we use [Hail](https://hail.is/) to merge the AoU and UKB variants.

Note that this work is part of a larger project to [Demonstrate the Potential for Pooled Analysis of All of Us and UK Biobank Genomic Data](https://github.com/all-of-us/ukb-cross-analysis-demo-project). Specifically this is for the portion of the project that is the **pooled** analysis.

# Setup 

<div class="alert alert-block alert-warning">
    <b>Cloud Environment</b>: This notebook was written for use on the <i>All of Us</i> Workbench.
    <ul>
        <li>Use "Recommended Environment" <kbd><b>Hail Genomics Analysis</b></kbd> which creates compute type <kbd>Dataproc Cluster</kbd> with reasonable defaults for CPU, RAM, disk, and number of workers. If you like, you can increase the number of workers to make this job complete faster.</li>
        <li>This notebook can take many hours to run. Recommend that it is run in the background via <kbd>run_notebook_in_the_background</kbd>.</li>
        <ul>
            <li>chr1 - chr22 took 2 days and 20 hours with Hail version 0.2.74-0c3a74d12093 due to using only 10 workers for the first 36 hours and then increasing to 40 workers for the remainder of the time. Recommend using a larger number of workers when running this in the future.</li>
            <li>We found that it was best to not use preemptible workers for this particular Hail job. It did not well-tolerate the failures due to preemptions.</li>
            <li>After our full run, we subsequently removed some code that was unnecessary, so this job may run faster in the future and possibly work just fine with some percentage of preemptible workers.</li>
        </ul>
    </ul>
</div>

from datetime import datetime
import hail as hl
import os
import time

## Retrieve the exome capture regions

For details, see https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=3803.

%%bash

wget -nd -nv biobank.ndph.ox.ac.uk/ukb/ukb/auxdata/xgen_plus_spikein.GRCh38.bed

gsutil cp xgen_plus_spikein.GRCh38.bed ${WORKSPACE_BUCKET}/data/ukb/

## Define constants

# Papermill parameters. See https://papermill.readthedocs.io/en/latest/usage-parameterize.html

#---[ Inputs ]---
# Matrix table was provided by AoU.
AOU_MT = 'gs://fc-secure-4adb21f6-46f4-4a79-99f9-afd63890c6d0/data/beta/beta_wgs_98622.mt'
# Matrix table was created from UKB 200k exome release VCFs.
# Note: The UKB matrix table was created via notebook 'aou_workbench_pooled_analyses/matrix_table_creation/create_matrix_tables.ipynb'
# and then repartitioned via notebook 'aou_workbench_pooled_analyses/matrix_table_creation/redo_partitions'.
UKB_MT = 'gs://fc-secure-fd6786bf-6c28-4f33-ac30-3860fbeee5bb/data/ukb/exomes/full_dataset_fewer_partitions.mt'
# Partition the two matrix tables with the same NUM_MT_READ_PARTITIONS partitions from the UKB matrix table.
# See also https://discuss.hail.is/t/improving-pipeline-performance/1344
NUM_MT_READ_PARTITIONS = 1250  # This number might be too low. Consider trying a higher number next time.
# This file is from https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=3803.
EXOME_REGIONS = f'{os.getenv("WORKSPACE_BUCKET")}/data/ukb/xgen_plus_spikein.GRCh38.bed'
# Use autosomes only.
INTERVALS_TO_EXAMINE = ['chr1-chr22']
INTERVALS_TO_EXAMINE_NAME = '_'.join(INTERVALS_TO_EXAMINE).replace(':', 'range')

#---[ Outputs ]---
# Create a timestamp for a folder of results generated today.
DATESTAMP = time.strftime('%Y%m%d')
TIMESTAMP = time.strftime('%Y%m%d_%H%M%S')
WORK_DIR = !pwd

AOU_ROWS_TAB = f'{os.getenv("WORKSPACE_BUCKET")}/data/pooled/geno/{DATESTAMP}/aou3-rows-{INTERVALS_TO_EXAMINE_NAME}.tab'
UKB_ROWS_TAB = f'{os.getenv("WORKSPACE_BUCKET")}/data/pooled/geno/{DATESTAMP}/ukb-rows-{INTERVALS_TO_EXAMINE_NAME}.tab'
MERGED_MT = f'{os.getenv("WORKSPACE_BUCKET")}/data/pooled/geno/{DATESTAMP}/merged-aou3-ukb-filtered-{INTERVALS_TO_EXAMINE_NAME}.mt'
AOU_ONLY_TAB = f'{os.getenv("WORKSPACE_BUCKET")}/data/pooled/geno/{DATESTAMP}/aou3-without_ukb-filtered-{INTERVALS_TO_EXAMINE_NAME}.tab'
UKB_ONLY_TAB = f'{os.getenv("WORKSPACE_BUCKET")}/data/pooled/geno/{DATESTAMP}/ukb-without_aou3-filtered-{INTERVALS_TO_EXAMINE_NAME}.tab'
HAIL_LOG = f'{WORK_DIR[0]}/hail-merge-variants-{TIMESTAMP}.log'
HAIL_LOG_DIR_FOR_PROVENANCE = f'{os.getenv("WORKSPACE_BUCKET")}/hail-logs/{DATESTAMP}/'

print(MERGED_MT)
print(AOU_ONLY_TAB)
print(UKB_ONLY_TAB)
print(HAIL_LOG)

## Check access

!gsutil ls {AOU_MT}

!gsutil ls {UKB_MT}

## Start Hail 

# See also https://towardsdatascience.com/fetch-failed-exception-in-apache-spark-decrypting-the-most-common-causes-b8dff21075c
# See https://spark.apache.org/docs/2.4.7/configuration.html

EXTRA_SPARK_CONFIG = {
    # If set to "true", performs speculative execution of tasks. This means if one or more tasks are running
    # slowly in a stage, they will be re-launched.
    'spark.speculation': 'true', # Default is false.
    
    # Fraction of tasks which must be complete before speculation is enabled for a particular stage.
    'spark.speculation.quantile': '0.95', # Default is 0.75

    # Default timeout for all network interactions. This config will be used in place of 
    # spark.core.connection.ack.wait.timeout, spark.storage.blockManagerSlaveTimeoutMs, 
    # spark.shuffle.io.connectionTimeout, spark.rpc.askTimeout or spark.rpc.lookupTimeout if they are not configured.
    'spark.network.timeout': '180s', # Default is 120s
        
    # (Netty only) Fetches that fail due to IO-related exceptions are automatically retried if this is set to a
    # non-zero value. This retry logic helps stabilize large shuffles in the face of long GC pauses or transient
    # network connectivity issues.
    'spark.shuffle.io.maxRetries': '10',  # Default is 3
    
    # (Netty only) How long to wait between retries of fetches. The maximum delay caused by retrying is 15 seconds
    # by default, calculated as maxRetries * retryWait.
    'spark.shuffle.io.retryWait': '15s',  # Default is 5s
    
    # Number of failures of any particular task before giving up on the job. The total number of failures spread
    # across different tasks will not cause the job to fail; a particular task has to fail this number of attempts.
    # Should be greater than or equal to 1. Number of allowed retries = this value - 1.
    'spark.task.maxFailures': '10', # Default is 4.

    # Number of consecutive stage attempts allowed before a stage is aborted.
    'spark.stage.maxConsecutiveAttempts': '10' # Default is 4.
}

hl.init(spark_conf=EXTRA_SPARK_CONFIG,
        min_block_size=50,
        default_reference='GRCh38',
        log=HAIL_LOG)

Check the configuration.

sc = hl.spark_context()
config = sc._conf.getAll()
config.sort()
config

start_all = datetime.now()
print(start_all)

# Load exome capture regions

ukb_exome_capture_regions = hl.import_bed(EXOME_REGIONS)

ukb_exome_capture_regions.describe()

ukb_exome_capture_regions.aggregate(hl.agg.counter(ukb_exome_capture_regions.interval.start.contig))

ukb_exome_capture_regions.show(5)

# Read UKB exomes matrix table

ukb_exomes = hl.read_matrix_table(UKB_MT)

ukb_exomes.n_partitions()

# https://discuss.hail.is/t/improving-pipeline-performance/1344
new_partitions = ukb_exomes._calculate_new_partitions(NUM_MT_READ_PARTITIONS)

ukb_exomes = hl.read_matrix_table(UKB_MT, _intervals=new_partitions)

ukb_exomes.describe()

## Filter to include only our genomic intervals of interest

INTERVALS_TO_EXAMINE

ukb_exomes = hl.filter_intervals(
    ukb_exomes,
    [hl.parse_locus_interval(x) for x in INTERVALS_TO_EXAMINE],
    keep=True)

## Filter to include only exonic variants

# Make the shuffle happen all at once by materializing the relevant row keys to a file.
relevant_ukb_rows = ukb_exomes.filter_rows(
    hl.is_defined(ukb_exome_capture_regions[ukb_exomes.locus])).rows().select().checkpoint(UKB_ROWS_TAB, overwrite=True)
ukb_exomes = ukb_exomes.semi_join_rows(relevant_ukb_rows) # alias for filter_rows(hl.is_defined ... )

# Read AoU matrix table

aou_wgs = hl.read_matrix_table(AOU_MT, _intervals=new_partitions)

aou_wgs.n_partitions()

aou_wgs.describe()

## Filter to include only our genomic intervals of interest

aou_wgs = hl.filter_intervals(
    aou_wgs,
    [hl.parse_locus_interval(x) for x in INTERVALS_TO_EXAMINE],
    keep=True)

## Filter to include only exonic variants

# Make the shuffle happen all at once by materializing the relevant row keys to a file.
relevant_aou_rows = aou_wgs.filter_rows(
    hl.is_defined(ukb_exome_capture_regions[aou_wgs.locus]), keep=True
  ).rows().select().checkpoint(AOU_ROWS_TAB, overwrite=True)
aou_wgs = aou_wgs.semi_join_rows(relevant_aou_rows) # alias for filter_rows(hl.is_defined ... )

# Examine variants with filter flags

<div class="alert alert-block alert-info">
<b>Note:</b> Note that AoU VCFs and UKB VCFs use different soft thresholds for what is flagged by VCF filters, so all variant QC happens further downstream.
</div>

aou_wgs_rows = aou_wgs.rows()
aou_wgs_rows.group_by(aou_wgs_rows.filters).aggregate(n = hl.agg.count()).show()

ukb_exomes_rows = ukb_exomes.rows()
ukb_exomes_rows.group_by(ukb_exomes_rows.filters).aggregate(n = hl.agg.count()).show()

# Perform the merge

## Split the multi-allelic sites

See also https://hail.is/docs/0.2/methods/genetics.html#hail.methods.split_multi_hts

# Split the multi-allelic sites into biallelic sites.
aou_prepared = hl.split_multi_hts(aou_wgs,
                                  keep_star=False,
                                  left_aligned=False,
                                  vep_root='vep',
                                  permit_shuffle=False)

# There are some ids found in both AoU and UKB, append a cohort suffix to differentiate them.
aou_prepared = aou_prepared.annotate_cols(cohort='AOU')
aou_prepared = aou_prepared.key_cols_by(aou_prepared.s, aou_prepared.cohort)

# Include only the row and entry fields that are needed.
aou_prepared = aou_prepared.select_entries(aou_prepared.GT)
aou_prepared = aou_prepared.select_rows(aou_qual=aou_prepared.qual,
                                        aou_filters=aou_prepared.filters,
                                        aou_info=aou_prepared.info,
                                        aou_a_index = aou_prepared.a_index,
                                        aou_was_split=aou_prepared.was_split,
                                       )

aou_prepared.describe()

# Split the multi-allelic sites into biallelic sites.
ukb_prepared = hl.split_multi_hts(ukb_exomes,
                                  keep_star=False,
                                  left_aligned=False,
                                  vep_root='vep',
                                  permit_shuffle=False)

# There are some ids found in both AoU and UKB, append a cohort suffix to differentiate them.
ukb_prepared = ukb_prepared.annotate_cols(cohort='UKB')
ukb_prepared = ukb_prepared.key_cols_by(ukb_prepared.s, ukb_prepared.cohort)

# Include only the row and entry fields that are needed.
ukb_prepared = ukb_prepared.select_entries(ukb_prepared.GT)
ukb_prepared = ukb_prepared.select_rows(ukb_qual=ukb_prepared.qual,
                                        ukb_filters=ukb_prepared.filters,
                                        ukb_info=ukb_prepared.info,
                                        ukb_a_index = ukb_prepared.a_index,
                                        ukb_was_split=ukb_prepared.was_split,
                                       )

ukb_prepared.describe()

## Compute the intersection

See also [union_cols](https://hail.is/docs/0.2/hail.MatrixTable.html#hail.MatrixTable.union_cols).

**Note:** The row fields from only one side of the merge can be retained. The AoU row fields contain much more information than those from UKB, so we retain the AoU row fields.

intersection = aou_prepared.union_cols(ukb_prepared)

# Uncomment to compute AC, AF and AN on the pooled cohort. Alternatively, plink2 can be used to compute these
# values on the resulting BGEN file exported from the matrix table.
#intersection = intersection.annotate_rows(gt_stats = hl.agg.call_stats(intersection.GT, intersection.alleles))

intersection.describe()

start = datetime.now()
print(start)

intersection.write(MERGED_MT, overwrite=True)

end = datetime.now()
print(end)
print(end - start)

## Compute AoU - UKB

The AoU data already contains AC, AF, and AN so we do not need to recompute it.

aou_only = aou_prepared.rows().anti_join(ukb_prepared.rows())

start = datetime.now()
print(start)

aou_only.write(AOU_ONLY_TAB, overwrite=True)

end = datetime.now()
print(end)
print(end - start)

## Compute UKB - AoU

ukb_only = ukb_prepared.rows().anti_join(aou_prepared.rows())

# Note that the UKB 200k exome data has fields for AC and AN but the values are all null. Comment the cell above and
# uncomment the lines below to compute AC, AF, and AN here while we still have the genotypes.
#ukb_only = ukb_prepared.anti_join_rows(aou_prepared.rows())
#ukb_only = ukb_only.annotate_rows(gt_stats = hl.agg.call_stats(ukb_only.GT, ukb_only.alleles))
#ukb_only = ukb_only.rows()

ukb_only.describe()

start = datetime.now()
print(start)

ukb_only.write(UKB_ONLY_TAB, overwrite=True)

end = datetime.now()
print(end)
print(end - start)

# Provenance

end_all = datetime.now()
print(end_all)
print(end_all - start_all)

# Copy the Hail log to the workspace bucket so that we can retain it.
!gzip --keep {HAIL_LOG}
!gsutil cp {HAIL_LOG}.gz {HAIL_LOG_DIR_FOR_PROVENANCE}

spark_history = [v for k, v in config if k == 'spark.history.fs.logDirectory']

spark_history

!gsutil ls {spark_history[0]}/**

print(datetime.now())

!pip3 freeze
